extract subject to its own module
